Return a 'not found' response for status requests with a malformed job id.
Job status requests containing malformed job ids (i.e. strings which do
not represent a valid UUID) return 500 Internal Server Error responses.
This is because UUID/fromString used to parse the id throws an exception
if the input cannot be converted to a UUID. Catch this exception and
return a 404 response if the job id is malformed.