Include extra information in API responses under a :details key.
The api-response function takes a map argument which is merged into
the default 'ok' response map which is {:type :ok}. If the given
map contains a :type key this value will overwrite the :ok value in
the map. Avoid this collision by appending the map to the response
document under a :details key. For example, a detail map of
{:type :other} will result in a response map of
{:type :ok :details {:type :other}}.